+++ src/ts/markdown/anchorRender.ts
@@ -1,5 +1,8 @@
-export const anchorRender = () => {
+export const anchorRender = (type: number) => {
     document.querySelectorAll(".vditor-anchor").forEach((anchor: HTMLLinkElement) => {
+        if (type === 1) {
+            anchor.classList.add("vditor-anchor--left");
+        }
         anchor.onclick = () => {
             const id = anchor.getAttribute("href").substr(1);
             const top = document.getElementById("vditorAnchor-" + id).offsetTop;