+++ CHANGELOG.md
@@ -67,6 +67,7 @@
 
 ### v3.2.4 / 2020-05-xx
 
+* [392](https://github.com/Vanessa219/vditor/issues/392) anchor option `改进功能`
 * [389](https://github.com/Vanessa219/vditor/issues/389) marker option at preview `改进功能`
 * [388](https://github.com/Vanessa219/vditor/pull/388) changed some korean i18n and demo text `文档相关`
 
@@ -94,6 +95,7 @@
   * 为 `options.upload` 添加 `extraData`
   * 添加静态方法 `mindmapRender`
   * 为 `IMarkdownConfig` 添加 `sanitize`, `listMarker` 配置
+  * IPreviewOptions.anchor 从 boolean 修改为 number
 
 ### v3.1.23 / 2020-05-05